Packages Restructured
All view elements have been organized into a new package called View with its subpackages
