+++ app/src/main/res/layout/activity_instagrammers.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".instagrammers.activity.InstagrammersListActivity">
+    tools:context=".instagrammers.view.activity.InstagrammersListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_height="wrap_content"
